# docs/swagger.yaml
definitions:
...
  github_com_sahandPgr_car-selling-service_api_dto.FileResponse:
    properties:
      description:
        type: string
      directory:
        type: string
      id:
        type: integer
      mediaType:
        type: string
      name:
        type: string
    type: object
  github_com_sahandPgr_car-selling-service_api_dto.Filter:
    properties:
      filterType:
        description: FilterType specifies the category of filter (e.g., "date", "number").
        type: string
      from:
        description: From specifies the starting value for the filter range.
        type: string
      to:
        description: To specifies the ending value for the filter range.
        type: string
      type:
        description: Type specifies the type of filter (e.g., "range").
        type: string
    type: object
...
  github_com_sahandPgr_car-selling-service_api_dto.PagedList-github_com_sahandPgr_car-selling-service_api_dto_FileResponse:
    properties:
      hasNextPage:
        description: HasNextPage indicates whether there is a next page.
        type: boolean
      hasPerviousPage:
        description: HasPerviousPage indicates whether there is a previous page.
        type: boolean
      items:
        description: Items contains the list of items for the current page.
        items:
          $ref: '#/definitions/github_com_sahandPgr_car-selling-service_api_dto.FileResponse'
        type: array
      pageNumber:
        description: PageNumber is the current page number.
        type: integer
      totalPages:
        description: TotalPages is the total number of pages available.
        type: integer
      totalRows:
        description: TotalRows is the total number of rows available.
        type: integer
    type: object
  github_com_sahandPgr_car-selling-service_api_dto.PaginationInputWithFilter:
    properties:
      filter:
        additionalProperties:
          $ref: '#/definitions/github_com_sahandPgr_car-selling-service_api_dto.Filter'
        description: Filter contains a map of column IDs to their filter configurations.
        type: object
      pageNumber:
        description: PageNumber specifies the current page number.
        type: integer
      pageSize:
        description: PageSize specifies the number of items per page.
        type: integer
      sort:
        description: Sort contains a list of sorting options.
        items:
          $ref: '#/definitions/github_com_sahandPgr_car-selling-service_api_dto.Sort'
        type: array
    type: object
...
  github_com_sahandPgr_car-selling-service_api_dto.Sort:
    properties:
      col_id:
        description: ColId specifies the column to sort.
        type: string
      sort:
        description: Sort specifies the sorting direction (e.g., "asc" or "desc").
        type: string
    type: object
...
  github_com_sahandPgr_car-selling-service_api_dto.UpdateFileRequest:
    properties:
      description:
        type: string
    type: object
  github_com_sahandPgr_car-selling-service_api_helper.BaseHttpResponse:
    properties:
      error: {}
      result: {}
      'statusCode:':
        type: integer
      success:
        type: boolean
      validationErrors:
        items:
          $ref: '#/definitions/github_com_sahandPgr_car-selling-service_api_validations.ValidationError'
        type: array
    type: object
  github_com_sahandPgr_car-selling-service_api_validations.ValidationError:
    properties:
      field:
        type: string
      message:
        type: string
      param:
        type: string
      tag:
        type: string
      value:
        type: string
    type: object
info:
  contact: {}
paths:
...
  /v1/files/:
    post:
      consumes:
      - application/x-www-form-urlencoded
      description: Create a File
      parameters:
      - in: formData
        name: desciption
        required: true
        type: string
      - description: Create a files
        in: formData
        name: file
        required: true
        type: file
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            allOf:
            - $ref: '#/definitions/github_com_sahandPgr_car-selling-service_api_helper.BaseHttpResponse'
            - properties:
                result:
                  $ref: '#/definitions/github_com_sahandPgr_car-selling-service_api_dto.FileResponse'
              type: object
        "400":
          description: Failed
          schema:
            $ref: '#/definitions/github_com_sahandPgr_car-selling-service_api_helper.BaseHttpResponse'
        "409":
          description: Failed
          schema:
            $ref: '#/definitions/github_com_sahandPgr_car-selling-service_api_helper.BaseHttpResponse'
      security:
      - AuthBearer: []
      summary: Create a File
      tags:
      - Files
  /v1/files/{id}:
    delete:
      consumes:
      - application/json
      description: Delete a File
      parameters:
      - description: Delete a files
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: Success
          schema:
            $ref: '#/definitions/github_com_sahandPgr_car-selling-service_api_helper.BaseHttpResponse'
        "400":
          description: Failed
          schema:
            $ref: '#/definitions/github_com_sahandPgr_car-selling-service_api_helper.BaseHttpResponse'
        "409":
          description: Failed
          schema:
            $ref: '#/definitions/github_com_sahandPgr_car-selling-service_api_helper.BaseHttpResponse'
      security:
      - AuthBearer: []
      summary: Delete a File
      tags:
      - Files
    get:
      consumes:
      - application/json
      description: GetById a File
      parameters:
      - description: GetById a files
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: Success
          schema:
            allOf:
            - $ref: '#/definitions/github_com_sahandPgr_car-selling-service_api_helper.BaseHttpResponse'
            - properties:
                result:
                  $ref: '#/definitions/github_com_sahandPgr_car-selling-service_api_dto.FileResponse'
              type: object
        "400":
          description: Failed
          schema:
            $ref: '#/definitions/github_com_sahandPgr_car-selling-service_api_helper.BaseHttpResponse'
        "409":
          description: Failed
          schema:
            $ref: '#/definitions/github_com_sahandPgr_car-selling-service_api_helper.BaseHttpResponse'
      security:
      - AuthBearer: []
      summary: GetById a File
      tags:
      - Files
    put:
      consumes:
      - application/json
      description: Update a File
      parameters:
      - description: Update a files
        in: path
        name: id
        required: true
        type: integer
      - description: Update a files
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/github_com_sahandPgr_car-selling-service_api_dto.UpdateFileRequest'
      produces:
      - application/json
      responses:
        "200":
          description: Success
          schema:
            allOf:
            - $ref: '#/definitions/github_com_sahandPgr_car-selling-service_api_helper.BaseHttpResponse'
            - properties:
                result:
                  $ref: '#/definitions/github_com_sahandPgr_car-selling-service_api_dto.FileResponse'
              type: object
        "400":
          description: Failed
          schema:
            $ref: '#/definitions/github_com_sahandPgr_car-selling-service_api_helper.BaseHttpResponse'
        "409":
          description: Failed
          schema:
            $ref: '#/definitions/github_com_sahandPgr_car-selling-service_api_helper.BaseHttpResponse'
      security:
      - AuthBearer: []
      summary: Update a File
      tags:
      - Files
  /v1/files/get-by-filter:
    post:
      consumes:
      - application/json
      description: Get Files
      parameters:
      - description: Request
        in: body
        name: Request
        required: true
        schema:
          $ref: '#/definitions/github_com_sahandPgr_car-selling-service_api_dto.PaginationInputWithFilter'
      produces:
      - application/json
      responses:
        "200":
          description: Success
          schema:
            allOf:
            - $ref: '#/definitions/github_com_sahandPgr_car-selling-service_api_helper.BaseHttpResponse'
            - properties:
                result:
                  $ref: '#/definitions/github_com_sahandPgr_car-selling-service_api_dto.PagedList-github_com_sahandPgr_car-selling-service_api_dto_FileResponse'
              type: object
        "400":
          description: Failed
          schema:
            $ref: '#/definitions/github_com_sahandPgr_car-selling-service_api_helper.BaseHttpResponse'
        "409":
          description: Failed
          schema:
            $ref: '#/definitions/github_com_sahandPgr_car-selling-service_api_helper.BaseHttpResponse'
      security:
      - AuthBearer: []
      summary: Get Files
      tags:
      - Files
...
securityDefinitions:
  AuthBearer:
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"